formpilot/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pymupdf==1.23.21
paddlepaddle==2.5.2
paddleocr==2.7.0.3
unstructured==0.11.8
sentence-transformers==2.2.2
rapidfuzz==3.6.1
pillow==10.2.0
numpy==1.24.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
coverage==7.4.0
ruff==0.1.13
python-dotenv==1.0.0